requirements.txt
pandas
openpyxl
pydantic
jsonlines
python-dotenv
tqdm
PyMySQL
